Adjusted release version to 0.11.0. (#133)
XML schema generation changes:
- Added support for inclusion of allowed values in XML schema as enumerations.
- Reorganized XML schema generation to provide for greater control over simple and complex type generation.
- Fixed bug causing unwrapped fields to not be generated properly.

JSON schema generation changes:
- Added support for inclusion of allowed values in JSON schema as enumerations.
- Refactored property generation to use a new property class hierarchy, which now encapsulates generation methods with their Metaschema constructs (i.e. flags, fields, assemblies).
- Completed refactor of JSON Schema object schema generation for Metaschema definitions and data types.
- Fixed ConcurrentModificationException caused by lazy schema initialization.

Model changes:
- Adjusted AutoCloser implementation, renaming methods to be more sensible.
- Adjusted assembly model containers to better expose choice instances. This will help in the future when adding support for usnistgov/metaschema#228.
- Improved error handling for Metapath compile errors to indicate the Metaschema the error occured in.

Code reorganization and method naming cleanup.
Cleaned up code, removing unused methods, imports, etc.
Fixed compile, PMD, and Spotbugs warnings.
Adopted latest metaschema commit
Updated GHA plugins to latest
Refactored and removed unneeded methods in both XML and JSON Schema generation.

+++ metaschema-testing/src/main/java/gov/nist/secauto/metaschema/model/testing/AbstractTestSuite.java
@@ -86,6 +86,8 @@ public abstract class AbstractTestSuite {
   private static final Logger LOGGER = LogManager.getLogger(AbstractTestSuite.class);
   private static final MetaschemaLoader LOADER = new MetaschemaLoader();
 
+  private static final boolean DELETE_RESULTS_ON_EXIT = false;
+  
   @NonNull
   protected abstract Format getRequiredContentFormat();
 
@@ -171,7 +173,9 @@ public abstract class AbstractTestSuite {
     Path collectionGenerationPath;
     try {
       collectionGenerationPath = Files.createTempDirectory(generationPath, "collection-");
-      deleteCollectionOnExit(collectionGenerationPath);
+      if (DELETE_RESULTS_ON_EXIT) {
+        deleteCollectionOnExit(collectionGenerationPath);
+      }
     } catch (IOException ex) {
       throw new JUnitException("Unable to create collection temp directory", ex);
     }
